requirements.txt
keyboard
numpy==1.26.4
opencv-python
pillow
pyautogui
pywinauto